build: bundle nanomsg-0.4-beta
I think this works, but it's probably a very bad way of doing it. we
should probably build nanomsg as a static library instead in a descrete
step in nanomsg.setup and link against that.

also the configuration, like NN_HAVE_EVENTFD, are hardcoded which is
bad. the bundled version should also be optional.